1.support delete and update in flink batch mode 2.supoort drop|create|select view 3.add test cases

+++ lakesoul-flink/src/main/java/org/apache/flink/lakesoul/metadata/LakeSoulCatalog.java
@@ -122,12 +122,26 @@ public class LakeSoulCatalog implements Catalog {
 
     @Override
     public List<String> listTables(String databaseName) throws CatalogException {
-        return dbManager.listTableNamesByNamespace(databaseName);
+        List<TableInfo> tifs = dbManager.getTableInfosByNamespace(databaseName);
+        List<String> tableNames = new ArrayList<>(100);
+        for (TableInfo item : tifs) {
+            if (!FlinkUtil.isView(item)) {
+                tableNames.add(item.getTableName());
+            }
+        }
+        return tableNames;
     }
 
     @Override
-    public List<String> listViews(String s) throws CatalogException {
-        throw new CatalogException("not supported now");
+    public List<String> listViews(String databaseName) throws CatalogException {
+        List<TableInfo> tifs = dbManager.getTableInfosByNamespace(databaseName);
+        List<String> tableNames = new ArrayList<>(100);
+        for (TableInfo item : tifs) {
+            if (FlinkUtil.isView(item)) {
+                tableNames.add(item.getTableName());
+            }
+        }
+        return tableNames;
     }
 
     @Override
@@ -162,12 +176,15 @@ public class LakeSoulCatalog implements Catalog {
             dbManager.deleteShortTableName(tableInfo.getTableName(), tableName, tablePath.getDatabaseName());
             dbManager.deleteDataCommitInfo(tableId);
             dbManager.deletePartitionInfoByTableId(tableId);
-            Path path = new Path(tableInfo.getTablePath());
-            try {
-                path.getFileSystem().delete(path, true);
-            } catch (IOException e) {
-                throw new RuntimeException(e);
+            if(!FlinkUtil.isView(tableInfo)){
+                Path path = new Path(tableInfo.getTablePath());
+                try {
+                    path.getFileSystem().delete(path, true);
+                } catch (IOException e) {
+                    throw new RuntimeException(e);
+                }
             }
+
         } else {
             if (ignoreIfNotExists) {
                 return;
@@ -216,7 +233,6 @@ public class LakeSoulCatalog implements Catalog {
         } else {
             cdcColumn = Optional.empty();
         }
-
         // adding hash bucket options
         if (!primaryKeys.isEmpty()) {
             if (Integer.parseInt(tableOptions.getOrDefault(HASH_BUCKET_NUM.key(), "-1")) <= 0) {
@@ -224,22 +240,28 @@ public class LakeSoulCatalog implements Catalog {
                         "Valid integer value for hashBucketNum property must be set for table with primary key");
             }
         }
-
+        String tableId = TABLE_ID_PREFIX + UUID.randomUUID();
+        String qualifiedPath = "";
+        String sparkSchema = FlinkUtil.toSparkSchema(schema, cdcColumn).json();
+        List<String> partitionKeys = Collections.emptyList();
+        if (table instanceof ResolvedCatalogTable) {
+            partitionKeys = ((ResolvedCatalogTable) table).getPartitionKeys();
+            String path = tableOptions.get(TABLE_PATH);
+            try {
+                FileSystem fileSystem = new Path(path).getFileSystem();
+                qualifiedPath = new Path(path).makeQualified(fileSystem).toString();
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
+        }
+        if (table instanceof ResolvedCatalogView) {
+            tableOptions.put(LAKESOUL_VIEW.key(), "true");
+            tableOptions.put(VIEW_ORIGINAL_QUERY,((ResolvedCatalogView) table).getOriginalQuery());
+            tableOptions.put(VIEW_EXPANDED_QUERY,((ResolvedCatalogView) table).getExpandedQuery());
+        }
         String json = JSON.toJSONString(tableOptions);
         JSONObject properties = JSON.parseObject(json);
-        List<String> partitionKeys = ((ResolvedCatalogTable) table).getPartitionKeys();
         String tableName = tablePath.getObjectName();
-        String path = tableOptions.get(TABLE_PATH);
-        String qualifiedPath = "";
-        try {
-            FileSystem fileSystem = new Path(path).getFileSystem();
-            qualifiedPath = new Path(path).makeQualified(fileSystem).toString();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-        String tableId = TABLE_ID_PREFIX + UUID.randomUUID();
-
-        String sparkSchema = FlinkUtil.toSparkSchema(schema, cdcColumn).json();
         dbManager.createNewTable(tableId, tablePath.getDatabaseName(), tableName, qualifiedPath, sparkSchema,
                 properties, DBUtil.formatTableInfoPartitionsField(primaryKeys, partitionKeys));
     }
